read/write to http.Requests
